#74 fix: PostActionMenu 공통 컴포넌트 추출 및 모바일 액션 메뉴 크래시 수정
